pyproject.toml
[project]
name = "stt-benchmark"
version = "0.1.0"
description = "STT (Speech-to-Text) TTFB benchmarking tool using Pipecat"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["stt", "speech-to-text", "benchmark", "ttfb", "pipecat", "latency"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # Pipecat with STT providers
    "pipecat-ai[assemblyai,aws,azure,cartesia,deepgram,elevenlabs,fal,gladia,google,gradium,groq,nvidia,openai,sambanova,sarvam,soniox,speechmatics,whisper,silero,soundfile]",
    # Dataset handling
    "datasets>=2.16.0,<3.0.0",
    "scipy>=1.11.0",
    # Configuration
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Storage
    "aiosqlite>=0.19.0",
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    # Utilities
    "loguru>=0.7.0",
    "numpy>=1.26.0",
    "aiohttp>=3.9.0",
    "librosa",
    # Semantic WER evaluation (Claude)
    "anthropic>=0.76.0",
    # Ground truth generation (Gemini)
    "google-genai>=0.3.0",
    # Interactive review (keyboard input)
    "readchar>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
stt-benchmark = "stt_benchmark.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stt_benchmark"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

 [tool.uv.sources]
pipecat-ai = { path = "../pipecat", editable = true }